+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 * Fix for `Array.sort` to make comparator argument optional
 * Tighten callback and predicate argument types for many Array methods: `every`, `filter`, `find`, `foreach`, ...
 * Internal-only fix to satisfy type checker variant used by the roblox-cli `convert` command
+* `Array.join` will now call `tostring()` on items before adding them
+
+### Added Polyfills
+* `String.charCodeAt`, `String.findOr`
+* `Array.includes`, `Array.unshift`
 
 ## 0.2.4